# Conflicts:
#	configure.in
#	doc/src/mainpage.dox
#	doc/src/tutorial/querying_devices.dox
#	doc/src/tutorial/tutorial_start.dox
#	src/hostapi/alsa/pa_linux_alsa.c
#	src/hostapi/coreaudio/pa_mac_core_blocking.c
#	src/hostapi/wasapi/pa_win_wasapi.c
#	src/hostapi/wdmks/pa_win_wdmks.c
#	src/os/win/pa_win_wdmks_utils.c

+++ src/hostapi/wmme/pa_win_wmme.c
@@ -817,7 +817,8 @@ static PaError InitializeInputDeviceInfo( PaWinMmeHostApiRepresentation *winMmeH
     {
         /* Append I/O suffix to WAVE_MAPPER device. */
         deviceName = (char *)PaUtil_GroupAllocateMemory(
-                    winMmeHostApi->allocations, StrTLen( wic.szPname ) + 1 + sizeof(constInputMapperSuffix_) );
+                    winMmeHostApi->allocations,
+					(long) (StrTLen( wic.szPname ) + 1 + sizeof(constInputMapperSuffix_)) );
         if( !deviceName )
         {
             result = paInsufficientMemory;
@@ -829,7 +830,8 @@ static PaError InitializeInputDeviceInfo( PaWinMmeHostApiRepresentation *winMmeH
     else
     {
         deviceName = (char*)PaUtil_GroupAllocateMemory(
-                    winMmeHostApi->allocations, StrTLen( wic.szPname ) + 1 );
+                    winMmeHostApi->allocations, 
+					(long) (StrTLen( wic.szPname ) + 1) );
         if( !deviceName )
         {
             result = paInsufficientMemory;
@@ -995,7 +997,8 @@ static PaError InitializeOutputDeviceInfo( PaWinMmeHostApiRepresentation *winMme
     {
         /* Append I/O suffix to WAVE_MAPPER device. */
         deviceName = (char *)PaUtil_GroupAllocateMemory(
-                    winMmeHostApi->allocations, StrTLen( woc.szPname ) + 1 + sizeof(constOutputMapperSuffix_) );
+                    winMmeHostApi->allocations, 
+					(long) (StrTLen( woc.szPname ) + 1 + sizeof(constOutputMapperSuffix_)) );
         if( !deviceName )
         {
             result = paInsufficientMemory;
@@ -1007,7 +1010,8 @@ static PaError InitializeOutputDeviceInfo( PaWinMmeHostApiRepresentation *winMme
     else
     {
         deviceName = (char*)PaUtil_GroupAllocateMemory(
-                    winMmeHostApi->allocations, StrTLen( woc.szPname ) + 1 );
+                    winMmeHostApi->allocations, 
+					(long) (StrTLen( woc.szPname ) + 1) );
         if( !deviceName )
         {
             result = paInsufficientMemory;
